+++ .gitignore
@@ -4,6 +4,7 @@
 build/
 out/
 mybuild/
+[Dd]ebug/
 
 # Files
 *.obj
