+++ CMakeUserPresets.json
@@ -1,0 +1,10 @@
+{
+    "version": 4,
+    "vendor": {
+        "conan": {}
+    },
+    "include": [
+        "build\\build\\generators\\CMakePresets.json",
+        "build\\generators\\CMakePresets.json"
+    ]
+}
