+++ CppLib/include/CConfigura.h
