+++ CppLib/include/CConfiguracion.h
