+++ CppLib/include/CLector.h
