+++ CppLib/include/COracle.h
