+++ CppLib/include/CParametro.h
