+++ CppLib/include/CRegistroLec.h
