+++ CppLib/include/CUtiles.h
