+++ CppLib/include/src/CConfigura.cpp
