+++ CppLib/include/src/CConfiguracion.cpp
