+++ CppLib/include/src/CLector.cpp
