+++ CppLib/include/src/CParametro.cpp
