+++ CppLib/include/src/CRegistroLec.cpp
